reformat code in single pages and fixed problems and tests

+++ resources/views/livewire/pages/animals/⚡index.blade.php
@@ -245,6 +245,7 @@ class extends Component {
     public function deleteAnimal(int $animalId): void
     {
         $animal = Animal::findOrFail($animalId);
+        $this->authorize('delete', $animal);
         $animal->delete();
     }
